requirements.txt
flask
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pymysql
mysql-connector-python
Flask-SQLAlchemy